+++ plugin-dev/Source/Sentry/Private/SentrySettings.cpp
@@ -45,6 +45,8 @@ USentrySettings::USentrySettings(const FObjectInitializer& ObjectInitializer)
 	, DiagnosticLevel(ESentryCliLogLevel::Info)
 	, UseLegacyGradlePlugin(false)
 	, CrashReporterUrl()
+	, bRequireUserConsent(false)
+	, bDefaultUserConsentGiven(true)
 	, bIsDirty(false)
 {
 	if (GIsEditor)